+++ src/codegen/builtin.rs
@@ -20,6 +20,7 @@ impl<'ctx> CodeGen<'ctx> {
             "@slice" => self.compile_slice(&args[0], &args[1], ret_ty),
             "@len" => self.compile_len(&args[0], ret_ty),
             "@inttoptr" => self.compile_inttoptr(&args[0], &args[1]),
+            "@ptrtoint" => self.compile_ptrtoint(&args[0], &args[1]),
             _ => unreachable!(),
         }
     }
@@ -118,6 +119,21 @@ impl<'ctx> CodeGen<'ctx> {
         Ok(Some(
             self.builder
                 .build_int_to_ptr(value.into_int_value(), ty, "int2ptr")
+                .as_basic_value_enum(),
+        ))
+    }
+
+    fn compile_ptrtoint(&mut self, ref_val: &AstNode, ty: &AstNode) -> CompRet<'ctx> {
+        let ref_val = get_value_from_result(&self.compile_node(ref_val)?)?;
+        let ty = match ty {
+            AstNode::Type(t) => self.okta_type_to_llvm(t),
+            _ => unreachable!(),
+        }
+        .into_int_type();
+
+        Ok(Some(
+            self.builder
+                .build_ptr_to_int(ref_val.into_pointer_value(), ty, "ptr2int")
                 .as_basic_value_enum(),
         ))
     }