+++ CHANGELOG.md
@@ -16,7 +16,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Support for path resolving and absolute paths when importing modules with `use`.
 - Language reference (`docs/`).
 - Compilation options from code ([#4](https://todo.sr.ht/~mikelma/oktac/4)).
-- `@inttoptr` built-in function.
+- `@inttoptr` and `@ptrtoint` built-in functions.
 
 ### Changed